chore: More tap refactoring
