+++ tap/package.json
@@ -7,7 +7,8 @@
     "build": "node build.mjs",
     "cbuild": "docker exec -it morio-tap node build.mjs",
     "clean": "rimraf dist",
-    "dev": "nodemon src/index.mjs -w dist --quiet -- --no-warnings",
+    "dev": "nodemon src/index.mjs -w src -- --no-warnings",
+    "test": "node src/index.mjs",
     "lint": "eslint --fix 'src/**'",
     "prettier": "npx prettier --ignore-unknown --write ./src"
   },
